Refactor auth middleware
